pyproject.toml
[tool.poetry]
name = "literarium"
version = "0.1.0"
description = ""
authors = ["Cristiano Tobias <cristiano.tobias40@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
djangorestframework = "^3.15.0"
django-rest-framework = "^0.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
factory-boy = "^3.3.0"
pytest-django = "^4.8.0"
faker = "^24.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
